+++ crates/outlayer/sdk/Cargo.toml
@@ -12,9 +12,14 @@ workspace = true
 crate-type = ["rlib"]
 
 [features]
+default = ["ed25519", "secp256k1"]
+
 ed25519 = ["defuse-outlayer-host/ed25519"]
 secp256k1 = ["defuse-outlayer-host/secp256k1"]
 
 [dependencies]
-defuse-outlayer-host = { workspace = true }
-wit-bindgen.workspace = true
+defuse-outlayer-host.workspace = true
+defuse-outlayer-sys.workspace = true
+
+# [target.'cfg(target_family = "wasm")'.dependencies]
+# defuse-outlayer-sys = { workspace = true }
